refactor(TALK): 채팅창 부분 DIR 분리
