+++ src/chat/talk/chatWindow.tsx
@@ -1,7 +1,7 @@
 import React from 'react';
 import { Socket } from 'socket.io-client';
 
-import ChatRoom from './chatRoom';
+import ChatRoom from './chatRoom/chatRoom';
 import SendForm from './sendForm';
 
 interface ChatWindowProps {
